chore(mock): Converting mock's Logger to use ListBase
It was using noSuchMethod and mirrors, both of which have the potential to bloat dart2js code size. This change may not have any real-world impact given the number of other places which also hit mirrors and NSM, but it seems good to remove the gratuitous use of these APIs.