Removing use of ClassMirror.members
This was removed from the public API prior to Dart 1.0 per https://groups.google.com/a/dartlang.org/forum/#!searchin/misc/members$20AND$20mirrors/misc/-BW-sAu_Us8/Nzh7OxOntB4J, but remained in the implementation.

I believe that the rest of the uses of this have been fixed up, this is the last item which causes Angular tests to fail on bleeding_edge Dart.