refactor(forms): remove _NgModelValidator in favor of using a shared interface
_NgModelValidator acted as the base class for validators in AngularDart with earlier
releases. This works, but it poses a problem for custom validators that exist outside
of the AngularDart core. This is because Inheritance is a problem since it is highly
limited when it comes to further refactoring and customization. Interfaces are more
versatile. Therefore all core and custom validators must implement the shared
interface so that they can be attached to ngModel for validation purposes.