fix(parser): remove map-specific code from the parser
The parser utility functions getKeyed() and setKeyed() treat instances
of the List and Map classes specially. For maps, the existing code
assumes that keys are always strings. If that is not the case, an
exception is likely thrown by the Map implementation. As map keys are
not necessarily strings, and there is code that handles generic
instances just fine, this commit simply removes the code that treats
maps specially.

Closes #608