fix(ng_control): Remove control from parent's errors when it is removed
At the moment, when a control is removed from its parent, the parent removes it from this list of
controls, but not from the list of errors. Thus, if a control is invalid when it is removed, the
parent does not currently update its validity. Change NgControl to also remove the child control
from its list of errors.