Serializing execution of transformers
When executing under pub serve the order of apply for the transformers may not be the same as the phase ordering. This was causing deadlocks for the transformers sharing a resolver as the read of an asset for one transformer was invoking the apply for a different transformer, which would then deadlock waiting for the resolver.