Fix for crashes in metadata extraction while in pub serve
Issue was around how the metadata extractor folds member annotations into the class annotations. It was doing in-place modifications to the AST which would then persist to the next build of the application in pub serve. Correct change is to clone the AST node which is to be modified and leave the original intact.