feat(NodeAttrs): Node attributes don't get initialized by NodeAttrs when setting up observing.
In case an element does not have a value set for a propertyv, that
property is not automatically initialised. For example:

  <div test foo="bar">

will cause ElementBinder to register observer (which will in turn cause
initialisation) on property foo (of the directive) to bar, but if we try
using

  <div test>

foo property will not be initialised.