feat: add option to use external css rewriter in template_cache_generator.
The provided rewriter will be used to transform every css read by the
TemplateCache, the original css is provided on stdin, and result is read
from the stdout of provided binary.

TESTING:

The test uses a simple sed script to exercise the integration.

The test runs only if sed is present on the running system.  The test
uses CSS with simple rules that are being translated for RTL locales.

Closes #1052