chore(intl): support newer version of "intl"
Angular's pubspec for intl spans versions with breaking changes.  For
Angular, a change in date formatting would be a breaking change but
changes do to CLDR updates are not breaking changes.  This commit
updates a test to swap out a locale with a CLDR update to one that
wasn't updated to enable it to pass across the CLDR breaking change.

Closes #1203