chore(travis): fail jobs when e2e tests fail
Note that this commit is actually going to cause Travis to fail since
the E2E tests are currently broken by passing due to issue #1528.  Once
this commit is in, I will submit #1527 to fix it.  These two commits
will serve as the test that this commit is actually doing the right
thing.

Closes #1528