fix(resource_url_resolver): IE/Safari compatible HTML parsing
On Internet Explorer (e.g. IE10 on Win7), the following weird bug occurs
with template elements (which are not supported, but still behave
differently than unknown tags.)

    var doc = new DOMParser().parseFromString("<!doctype html><html><body><template>CONTENTS</template></body></html>", "text/html");

    // Prints "<template>CONTENTS</template>" as expected.
    console.log(doc.body.innerHTML);

    // This should be a no-op.
    doc.body.querySelectorAll("div");

    // Prints "<template></template>" - completely losing "CONTENTS".
    console.log(doc.body.innerHTML);

In addition, Safari's DOMParser parses XML but not HTML, so there was
already an alternate path for it.

In this commit, I use a single compatible path for all these browsers.