fix(resource_url_resolver): Makes resolved URLs absolute
Makes resolved type URLs absolute instead of relative. When they are relative
it is difficult to reuse components. Using pub serve relative URLs are resolved
because pub does some magic to allow add symlinks to the packages directory
where needed. In production relative URLs make it difficult to add file handlers
to the locations that are needed, as the location will be determined based on
the current browser path or route. It also makes the template cache harder to
populate because the same resource may need to be served from two separate
URLs.

Closes #1599