<fix>(transformers): Fix relative URL lookup for transformers, and web
Extracts logic used in the TypeRelativeUriGenerator to be used in other
transformers. This allows other transformers such as the ExpressionGenerator
and later the TemplateCacheGenerator to find relative resources.

Fixes asset resolution according to Barback rules which are defined here:
http://goo.gl/YDMRc2 specifically library assets are defined with packages
being in the URI and all other resources are considered web resources.

Closes #1649, #1651